Refactor: Split OTP verification logic
Extract OTP verification into a separate component for better maintainability.
